requirements.txt
pandas
numpy
matplotlib
yfinance
pyyaml
